% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bayesian.R
\name{.resolve_stan_model}
\alias{.resolve_stan_model}
\title{Resolve Stan model file path}
\usage{
.resolve_stan_model(stan_file, stan_code = NULL, package_name = NULL)
}
\description{
Resolve Stan model file path
}
\keyword{internal}
