% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_covariate_effects}
\alias{plot_covariate_effects}
\title{Plot Covariate Coefficients (Forest Plot)}
\usage{
plot_covariate_effects(fit, stan_data)
}
\arguments{
\item{fit}{A \code{stanfit} object.}

\item{stan_data}{The list data passed to Stan.}
}
\value{
A ggplot object.
}
\description{
Creates a forest plot showing the effects of covariates on susceptibility
and infectivity.
}
\keyword{internal}
