% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_posterior_distributions}
\alias{plot_posterior_distributions}
\title{Plot Posterior Distributions (Phi and Kappa)}
\usage{
plot_posterior_distributions(fit)
}
\arguments{
\item{fit}{A \code{stanfit} object.}
}
\value{
A combined ggplot object.
}
\description{
Creates violin plots of posterior distributions for susceptibility and
infectivity parameters by role.
}
\keyword{internal}
