% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransmissionChainAnalysis.R
\name{print.TransmissionChainResult}
\alias{print.TransmissionChainResult}
\title{Print a TransmissionChainResult}
\usage{
\method{print}{TransmissionChainResult}(x, ...)
}
\arguments{
\item{x}{A \code{TransmissionChainResult} object.}

\item{...}{Passed to or from other methods (unused).}
}
\value{
\code{x}, returned invisibly.
}
\description{
Nicely prints sections available in a \code{TransmissionChainResult} returned by
\code{\link{TransmissionChainAnalysis}}.
}
\seealso{
\code{\link{TransmissionChainAnalysis}}
}
