% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{summarize_attack_rates}
\alias{summarize_attack_rates}
\title{Summarize Attack Rates & Reinfections}
\usage{
summarize_attack_rates(sim_result)
}
\arguments{
\item{sim_result}{A list object returned by \code{simulate_multiple_households_comm}.}
}
\value{
A list containing four dataframes:
\describe{
\item{primary_overall}{Overall primary attack rate statistics.}
\item{primary_by_role}{Primary attack rate stratified by age group.}
\item{reinf_overall}{Overall summary of reinfection counts and rates.}
\item{reinf_by_role}{Reinfection counts stratified by age group.}
}
}
\description{
Calculates the Primary Attack Rate (proportion of people infected at least once)
and a separate summary of Reinfections (secondary episodes).
}
\examples{
\dontrun{
household_profile <- list(
  prob_adults   = c(0, 0, 1),
  prob_infant   = 1.0,
  prob_siblings = c(0, .8, .2),
  prob_elderly  = c(0.7, 0.1, 0.2)
)

sim_res <- simulate_multiple_households_comm(
  n_households = 100,
  viral_testing = "viral load",
  infectious_shape = 10,
  infectious_scale = 1,
  waning_shape = 6,
  waning_scale = 10,
  surveillance_interval = 4,
  start_date = “2024-07-01”,
  end_date = “2025-06-30”,
  surveillance_df = surveillance_data,
  seed = 123,
  household_profile_list = household_profile
)
rates <- summarize_attack_rates(sim_res)
}
}
