% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{fit_household_model}
\alias{fit_household_model}
\title{Fit Household Transmission Model}
\usage{
fit_household_model(
  stan_data,
  iter = 2000,
  chains = 4,
  warmup = 1000,
  init_fun = NULL,
  control = list(adapt_delta = 0.95, max_treedepth = 15),
  cores = 4,
  refresh = 50,
  ...
)
}
\arguments{
\item{stan_data}{A list of data formatted by \code{prepare_stan_data}.}

\item{iter}{Integer; number of iterations per chain (including warmup).}

\item{chains}{Integer; number of Markov chains.}

\item{warmup}{Integer; number of warmup iterations.}

\item{init_fun}{Function or List; initial values for the sampler.}

\item{control}{List; Stan control parameters.}

\item{cores}{Integer; number of CPU cores.}

\item{refresh}{Integer; refresh rate for output.}

\item{...}{Additional arguments passed to \code{rstan::sampling}.}
}
\value{
A \code{stanfit} object containing the posterior samples.
}
\description{
Fits the compiled Stan model to the prepared household data.
}
\examples{
\dontrun{
household_profile <- list(
  prob_adults   = c(0, 0, 1),
  prob_infant   = 1.0,
  prob_siblings = c(0, .8, .2),
  prob_elderly  = c(0.7, 0.1, 0.2)
)

sim_res <- simulate_multiple_households_comm(
  n_households = 100,
  viral_testing = "viral load",
  infectious_shape = 10,
  infectious_scale = 1,
  waning_shape = 6,
  waning_scale = 10,
  surveillance_interval = 4,
  start_date = “2024-07-01”,
  end_date = “2025-06-30”,
  surveillance_df = surveillance_data,
  seed = 123,
  household_profile_list = household_profile
)
person_covariates <- sim_res$hh_df \%>\%
  dplyr::select(hh_id, person_id, vacc_status) \%>\%   dplyr::distinct()

df_for_stan <- sim_res$diagnostic_df \%>\%
  dplyr::left_join(person_covariates, by = c("hh_id", "person_id"))

stan_input <- prepare_stan_data(
  df_clean = df_for_stan,
  use_vl_data      = 1,
  study_start_date = as.Date(study_start),
  study_end_date   = as.Date(study_end),
  surveillance_df  = surveillance_data,
  study_start_date = as.Date(study_start),
  study_end_date   = as.Date(study_end),
  imputation_params = VL_params_list)
fit <- fit_household_model(stan_input)
}
}
