% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSyn.R
\name{print.GenSynResult}
\alias{print.GenSynResult}
\title{Print method for \code{GenSynResult}}
\usage{
\method{print}{GenSynResult}(x, ...)
}
\arguments{
\item{x}{A \code{GenSynResult} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{GenSynResult}
}
